2. Cleaned some files

+++ dhis-2/dhis-web/dhis-web-dashboard-integration/src/main/webapp/dhis-web-dashboard-integration/javascript/dashboard.js
@@ -1,7 +1,25 @@
 
 dhis2.util.namespace( 'dhis2.db' );
 
-dhis2.db.current;
+dhis2.db.currentKey = "dhis2.dashboard.current";
+
+dhis2.db.current = function() 
+{
+	var current = localStorage[dhis2.db.currentKey];
+	
+	return current;
+}
+
+dhis2.db.setCurrent = function( id )
+{
+	localStorage[dhis2.db.currentKey] = id;
+}
+
+dhis2.db.clearCurrent = function()
+{
+	localStorage.removeItem( dhis2.db.currentKey );
+}
+
 dhis2.db.currentItem;
 dhis2.db.currentItemPos;
 dhis2.db.currentShareType;
@@ -110,13 +128,15 @@ dhis2.db.dashboardReady = function( id )
 	} );	
 }
 
-dhis2.db.dragStart = function( event, ui ) {
+dhis2.db.dragStart = function( event, ui ) 
+{
 	$( this ).hide();
 	dhis2.db.currentItem = $( this ).attr( "id" );
 	dhis2.db.currentItemPos = dhis2.db.getIndex( dhis2.db.currentItem );
 }
 
-dhis2.db.dragStop = function( event, ui ) {
+dhis2.db.dragStop = function( event, ui ) 
+{
 	$( this ).show();
 	$( ".dropItem" ).not( ".lastDropItem" ).hide();
 	$( ".lastDropItem" ).removeClass( "blankDropItem" ).addClass( "blankDropItem" );
@@ -124,19 +144,22 @@ dhis2.db.dragStop = function( event, ui ) {
 	dhis2.db.currentItemPos = undefined;
 }
 
-dhis2.db.dropOver = function( event, ui ) {
+dhis2.db.dropOver = function( event, ui ) 
+{
 	var itemId = $( this ).attr( "id" );
 	var dropItemId = "drop-" + itemId;
 	$( "#" + dropItemId ).show();
 }
 
-dhis2.db.dropOut = function( event, ui ) {
+dhis2.db.dropOut = function( event, ui ) 
+{
 	var itemId = $( this ).attr( "id" );
 	var dropItemId = "drop-" + itemId;
 	$( "#" + dropItemId ).hide();
 }
 
-dhis2.db.lastDropOver = function( event, ui ) {
+dhis2.db.lastDropOver = function( event, ui ) 
+{
 	$( this ).removeClass( "blankDropItem" ).css( "display", "block" );
 }
 
@@ -179,9 +202,9 @@ dhis2.db.openAddDashboardForm = function()
 
 dhis2.db.openManageDashboardForm = function()
 {
-	if ( undefined !== dhis2.db.current )
+	if ( undefined !== dhis2.db.current() )
 	{
-		$.getJSON( "../api/dashboards/" + dhis2.db.current, function( data )
+		$.getJSON( "../api/dashboards/" + dhis2.db.current(), function( data )
 		{
 			var name = data.name;
 			$( "#dashboardRename" ).val( name );
@@ -216,7 +239,7 @@ dhis2.db.addDashboard = function()
 			if ( location !== undefined && location.lastIndexOf( "/" ) != -1 )
 			{
 				var itemId = location.substring( ( location.lastIndexOf( "/" ) + 1 ) );
-				dhis2.db.current = itemId;				
+				dhis2.db.setCurrent( itemId );				
 			}
 			
 			dhis2.db.renderDashboardListLoadFirst();
@@ -230,13 +253,13 @@ dhis2.db.renameDashboard = function()
 	
 	$( "#manageDashboardForm" ).dialog( "destroy" );
 	
-	if ( undefined !== dhis2.db.current && undefined !== name && name.trim().length > 0 )
+	if ( undefined !== dhis2.db.current() && undefined !== name && name.trim().length > 0 )
 	{
 		var data = "{ \"name\": \"" + name + "\"}";
 		
 		$.ajax( {
 	    	type: "put",
-	    	url: "../api/dashboards/" + dhis2.db.current,
+	    	url: "../api/dashboards/" + dhis2.db.current(),
 	    	contentType: "application/json",
 	    	data: data,
 	    	success: function() {
@@ -249,21 +272,22 @@ dhis2.db.renameDashboard = function()
 
 dhis2.db.removeDashboard = function()
 {
-	if ( undefined !== dhis2.db.current )
+	if ( undefined !== dhis2.db.current() )
 	{
 		$.ajax( {
 			type: "delete",
-			url: "../api/dashboards/" + dhis2.db.current,
+			url: "../api/dashboards/" + dhis2.db.current(),
 	    	success: function() {
 	    		$( "#manageDashboardForm" ).dialog( "destroy" );
-	    		dhis2.db.current = undefined;
+	    		dhis2.db.clearCurrent();
 	    		dhis2.db.renderDashboardListLoadFirst();
 	    	}
 		} );
 	}
 }
 
-function updateSharing( dashboard ) {
+function updateSharing( dashboard ) 
+{
     $( '#dashboardMenu' ).data( 'canManage', dashboard.access.manage );
 
     if ( dashboard.access.manage ) {
@@ -325,12 +349,12 @@ dhis2.db.renderDashboardListLoadFirst = function()
 				}
 			} );
 
-			if ( undefined == dhis2.db.current )
+			if ( undefined === dhis2.db.current() )
 			{
-				dhis2.db.current = first;
+				dhis2.db.setCurrent( first );
 			}
 			
-			dhis2.db.renderDashboard( dhis2.db.current );		
+			dhis2.db.renderDashboard( dhis2.db.current() );		
 		}
 		else
 		{
@@ -347,11 +371,11 @@ dhis2.db.clearDashboard = function()
 
 dhis2.db.renderDashboard = function( id )
 {
-	$( "#dashboard-" + dhis2.db.current ).removeClass( "currentDashboard" );
+	$( "#dashboard-" + dhis2.db.current() ).removeClass( "currentDashboard" );
 	
-	dhis2.db.current = id;
+	dhis2.db.setCurrent( id );
 	
-	$( "#dashboard-" + dhis2.db.current ).addClass( "currentDashboard" );
+	$( "#dashboard-" + dhis2.db.current() ).addClass( "currentDashboard" );
 	
 	$.getJSON( "../api/dashboards/" + id, function( data )
 	{
@@ -362,7 +386,12 @@ dhis2.db.renderDashboard = function( id )
         if( undefined !== data.items )
         {
 			$.each( data.items, function( index, item )
-			{				
+			{
+				if ( null == item || undefined === item )
+				{
+					return true;
+				}
+				
 				if ( "chart" == item.type )
 				{
 					$d.append( $.tmpl( dhis2.db.tmpl.chartItem, { "itemId": item.id, "id": item.chart.id, "name": item.chart.name, 
@@ -455,6 +484,11 @@ dhis2.db.renderLinkItem = function( $d, itemId, contents, title, baseUrl, urlSuf
 	
 	$.each( contents, function( index, content )
 	{
+		if ( null == content || undefined === content )
+		{
+			return true;
+		}		
+		
 		html += 
 			"<li><a href='" + baseUrl + content.id + urlSuffix + "'>" + content.name + "</a><a class='removeItemLink' href='javascript:dhis2.db.removeItemContent( \"" + itemId + "\", \"" + content.id + "\" )' title='" + i18n_remove + "'>" + 
 			"<img src='../images/hide.png'></a></li>";
@@ -481,7 +515,7 @@ dhis2.db.moveItem = function( id, destItemId, position )
 	$targetDropLi.insertBefore( $destLi );
 	$targetLi.insertBefore( $destLi );
 		
-	var url = "../api/dashboards/" + dhis2.db.current + "/items/" + id + "/position/" + position;
+	var url = "../api/dashboards/" + dhis2.db.current() + "/items/" + id + "/position/" + position;
 	
 	$.post( url, function() {
 	} );
@@ -489,17 +523,17 @@ dhis2.db.moveItem = function( id, destItemId, position )
 
 dhis2.db.addItemContent = function( type, id )
 {
-	if ( undefined !== dhis2.db.current )
+	if ( undefined !== dhis2.db.current() )
 	{
 		$.ajax( {
 	    	type: "post",
-	    	url: "../api/dashboards/" + dhis2.db.current + "/items/content",
+	    	url: "../api/dashboards/" + dhis2.db.current() + "/items/content",
 	    	data: {
 	    		type: type,
 	    		id: id
 	    	},
 	    	success: function() {
-	    		dhis2.db.renderDashboard( dhis2.db.current );
+	    		dhis2.db.renderDashboard( dhis2.db.current() );
 	    	},
 	    	error: function( xhr ) {
 	    		setHeaderDelayMessage( xhr.responseText );
@@ -512,10 +546,10 @@ dhis2.db.removeItem = function( id )
 {
 	$.ajax( {
     	type: "delete",
-    	url: "../api/dashboards/" + dhis2.db.current + "/items/" + id,
+    	url: "../api/dashboards/" + dhis2.db.current() + "/items/" + id,
     	success: function() {
     		dhis2.db.currentItem = undefined;
-    		dhis2.db.renderDashboard( dhis2.db.current );
+    		dhis2.db.renderDashboard( dhis2.db.current() );
     	}
     } );
 }
@@ -524,9 +558,9 @@ dhis2.db.removeItemContent = function( itemId, contentId )
 {
 	$.ajax( {
     	type: "delete",
-    	url: "../api/dashboards/" + dhis2.db.current + "/items/" + itemId + "/content/" + contentId,
+    	url: "../api/dashboards/" + dhis2.db.current() + "/items/" + itemId + "/content/" + contentId,
     	success: function() {
-    		dhis2.db.renderDashboard( dhis2.db.current );
+    		dhis2.db.renderDashboard( dhis2.db.current() );
     	}
     } );
 }
